+++ kernel/server/view/block.go
@@ -15,6 +15,8 @@ func Block(c *gin.Context, doc *models.Doc) {
 		title_height string   // 题头最小高度
 		tags         []string // 文档标签
 	)
+
+	/* 文档题头 */
 	if doc.TitleImg != "" { // 有题头图
 		title_height = "30vh"
 	} else if doc.Icon != "" { // 无题头图, 有图标
@@ -23,11 +25,13 @@ func Block(c *gin.Context, doc *models.Doc) {
 		title_height = "0"
 	}
 
+	/* 文档标签 */
 	if doc.Tag != "" { // 有文档标签
 		tags = strings.Split(doc.Tag[1:len(doc.Tag)-1], "# #")
 	} else {
 		tags = []string{}
 	}
+
 	c.HTML(http.StatusOK, "block.html", gin.H{
 		"Path":        doc.Path,
 		"Hpath":       doc.Hpath,